backend/text-to-speech/requirements.txt

torchaudio
modal 
pydantic
fastapi
requests 
chatterbox-tts